tools/local_exp/requirements.txt
azure-ai-evaluation
dotenv